resume-analyzer/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9
jinja2>=3.1.4
pydantic>=2.7.0,<3.0.0
spacy>=3.7.0,<4.0.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.2
